+++ lib/PHPPdf/Core/Parser/DocumentParserListener.php
@@ -17,8 +17,8 @@ use PHPPdf\Core\Node\Node;
  */
 interface DocumentParserListener
 {
-    public function onStartParseNode(Document $document, PageCollection $root, Node $node);
-    public function onEndParseNode(Document $document, PageCollection $root, Node $node);
-    public function onEndParsePlaceholders(Document $document, PageCollection $root, Node $node);
+    public function onStartParseNode(Document $document, PageCollection $root, Node $node, DocumentParsingContext $context);
+    public function onEndParseNode(Document $document, PageCollection $root, Node $node, DocumentParsingContext $context);
+    public function onEndParsePlaceholders(Document $document, PageCollection $root, Node $node, DocumentParsingContext $context);
     public function onEndParsing(Document $document, PageCollection $root);
 }